+++ src/main/java/baseball/BaseballGame.java
@@ -2,19 +2,19 @@ package baseball;
 
 import baseball.console.GameConsole;
 import baseball.core.BaseballGameCore;
-import baseball.printer.GameMessagePrinter;
+import baseball.console.output.ConsoleOutput;
 
-import static baseball.console.converter.RestartOrExitCodeConverter.EXIT_CODE;
-import static baseball.console.converter.RestartOrExitCodeConverter.START_CODE;
+import static baseball.console.input.converter.RestartOrExitCodeConverter.EXIT_CODE;
+import static baseball.console.input.converter.RestartOrExitCodeConverter.START_CODE;
 
 public class BaseballGame {
 
-    private final GameMessagePrinter messagePrinter;
+    private final ConsoleOutput messagePrinter;
     private final GameConsole console;
     private final BaseballGameCore gameCore;
 
     public BaseballGame() {
-        this.messagePrinter = new GameMessagePrinter();
+        this.messagePrinter = new ConsoleOutput();
         this.console = new GameConsole();
         this.gameCore = new BaseballGameCore(messagePrinter, console);
     }
